(test) groups

+++ server/src/contexts/module_magazine/magazine/infrastructure/schemas/magazine.schema.ts
@@ -35,8 +35,11 @@ MagazineSchema.pre(
         sectionsToDelete.push(docs[i].sections[j])
       }
     }
-    await docs[0].model('MagazineSection').deleteMany({ _id: { $in: sectionsToDelete } });
 
+
+    if (sectionsToDelete.length > 0) {
+      await docs[0].model('MagazineSection').deleteMany({ _id: { $in: sectionsToDelete } });
+    }
     next();
   },
 );
